requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-limiter==3.5.0
flask-caching==2.1.0
flask-compress==1.14.0
python-dotenv==1.0.1
anthropic==0.18.1
python-json-logger==2.0.7
prometheus-client==0.20.0
firebase-admin==6.4.0
gunicorn==21.2.0
redis==5.0.1